ADD Test

+++ RPI/inc/main.h
@@ -72,7 +72,7 @@ extern PID   dc2;
 extern PID   dc3;
 extern PID   bl;
 extern SERVO servo;
-extern float bldcSpeed;
+extern int bldcSpeed;
 extern void setSeparatePID(PID *pid);
 
 #endif
